Add AbstractTracerTest which tests the ContextTracer and ExtensionTracer
